pyproject.toml
[build-system]
requires = ["setuptools>=40.8.0"]

[project]
name = "keykeeper"
version = "0.1.0"
dependencies = [
    "aiofiles==25.1.0",
    "aiosqlite==0.22.1",
    "aiohttp==3.13.3",
    "pycryptodome==3.23.0",
    "click==8.3.1"
]

requires-python = ">= 3.11"
authors = [
  { name="Roman Volkov", email="althusky@gmail.com" },
]
description = "Lightweight Secret Storage Server"
readme = "README.md"
license = "GPL-3.0-or-later"
license-files = ["LICENSE"]

[project.urls]
Homepage = "https://github.com/althusky/keykeeper"

[project.scripts]
keykeeperd = "keykeeper.server:main"
keykeeper = "keykeeper.keykeeper:cli"
